+++ tasks/options/jsify.js
@@ -4,19 +4,6 @@
     var grunt = require('grunt');
 
     module.exports = {
-        splash: {
-            options: {
-                transform: function(orig) {
-                    return 'module.exports = ' + orig + ';';
-                }
-            },
-            expand: true,
-            cwd: '.tmp/collateral/',
-            src: ['**/*.html'],
-            dest: '.tmp/collateral/',
-            ext: '.js',
-            extDot: 'last'
-        },
         dist: {
             options: {
                 transform: function(orig, srcs) {